feat(user-libs): add "achordion" to my keymap
Achordion is a user-space library that handles "bilateral combinations"
for tap-hold (home row) mods. I.e., it disables resolving a tap-hold key
as held, when followed by another key on the same side of the keyboard.

So in my case, `s` (`d` on qwerty), is an `s` when tapped, `shift` when
held, and `s` when succeeded by `t` (`f` on qwerty). This prevents false
positives when rolling keys on the same hand, and forces one to
alternate the hands they use for modifiers, which is good form in some
situations.

Additionally, Achordion offer a "streak timeout". When a mod-tap key is
pressed within this timeframe _after_ a regular key, it is resolved as a
tap. This helps prevent false-positives when typing fast. Though my top
typing speed hasn't increased much, I do think the experience has gotten
even better by a little bit.